[pt-br] Translation of the State of Rust Survey FAQ
There is a growing community in Brasil interested in Rust, and we would like to have their participation on the `State of Rust Survey`, allowing them to express on the language they feel most comfortable to participate.

This effort means we would be able to translate the `Survey` to Portuguese, giving the chance to have inputs of members of the community that feel more comfortable to write on their native tongue.
To make they welcome, it would be nice to have a translation of the FAQ, not only of the survey, to give provide details on how the data collected on the survey will be used.

This commit is the initial translation of the [English FAQ](https://github.com/rust-community/team/wiki/State-of-the-Rust-Language-Community-Survey-FAQ) on the Wiki. This document was put on the main repo to allow contributions in the form of Pull Requests.
The initial idea is to have the same file name under the `docs/pt-br` folder, to accommodate more translations if there is interest.